BZ 1064913 GFS2: Fix slab memory leak in gfs2_bufdata
This patch fixes a slab memory leak that sometimes can occur
for files with a very short lifespan. The problem occurs when
a dinode is deleted before it has gotten to the journal properly.
In the leak scenario, the bd object is pinned for journal
committment (queued to the metadata buffers queue: sd_log_le_buf)
but is subsequently unpinned and dequeued before it finds its way
to the ail or the revoke queue. In this rare circumstance, the bd
object needs to be freed from slab memory, or it is forgotten.
We have to be very careful how we do it, though, because
multiple processes can call gfs2_remove_from_journal. In order to
avoid double-frees, only the process that does the unpinning is
allowed to free the bd.

BZ 1064913 Use-after-free and memory leak in GFS2 when removing from journal.

Previously, the GFS2 kernel module leaked memory in the gfs2_bufdata slab
cache and allowed a use-after-free race condition to be triggered in the
gfs2_remove_from_journal() function. As a consequence after unmounting the
GFS2 file system, the GFS2 slab cache could still contain some objects,
which subsequently could, under certain circumstances, result in a kernel
panic. A series of patches has been applied to the GFS2 kernel module,
ensuring that all objects are freed from the slab cache properly and the
kernel panic is avoided.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>
(cherry picked from commit 502be2a32f09f388e4ff34ef2e3ebcabbbb261da)