BZ 1063507 Denial-of-service during boot if LAPIC is not initialized early.
A previous change in the Advanced Programmable Interrupt Controller (APIC)
code caused a regression on certain Intel CPUs using a Multiprocessor (MP)
table. An attempt to read from the local APIC (LAPIC) could be performed
before the LAPIC was mapped, resulting in a kernel crash during a system
boot. A patch has been applied to fix this problem by mapping the LAPIC as
soon as possible when parsing the MP table.