RHEL IPMI fixes.
A bug in the IPMI driver caused the kernel to panic when an IPMI
interface was removed using the hotmod script. The IPMI driver has been
fixed to properly clean the relevant data when removing an IPMI
interface.
(BZ#1149578)

Due to a bug in the IPMI driver, the kernel could panic when adding an
IPMI interface that was previously removed using the hotmod script.
This update fixes this bug by ensuring that the relevant shadow
structure
is initialized at the right time. (BZ#1149580)

Signed-off-by: Jamie Iles <jamie.iles@oracle.com>