RHEL boot failure with mgr_rxbuf=off boot parameter.
Due to previous changes in the virtio-net driver, a Red Hat Enterprise
Linux 6.6 guest was unable to boot with the "mgr_rxbuf=off" option
specified. This was caused by providing the page_to_skb() function with
an incorrect packet length in the driver's Rx path. This problem has
been fixed and the guest in the described scenario can now boot
successfully.  (BZ#1148693)

Signed-off-by: Jamie Iles <jamie.iles@oracle.com>